Update 2017-06-14-data-plumbing.md
Minor typos/grammatical errors corrected.

Generally-speaking, I think it's well-written, but at this point, the benefits (and examples of custom Avro) represent 50% of this total file.  Perhaps that was intentional (although, I suspect it wasn't, even if it's both "central" and "most-impactful" to automating the low-level ingestion work).  The (non-Avro) topics (by relative comparison) feel a bit "rushed".  I think it's better if either the full picture is provided (but eliminate some of the Avro-content providing examples to allow all tooling components to have relative, sufficient representation), or quickly list the others while highlighting that they will be explored more in future blog posts.